groups admin change and delete of a group
When the admin of a group tries to leave:
Show a popup dialog listing all group members (except the admin).
The admin must select a new admin from the list.
On confirmation, transfer admin rights to the selected member and then leave the group.
and added the option to delete a group